# Revert #26418, remove noinline annotation from fill!