@nanosoldier 2019-04-05
[Your benchmark job](https://github.com/JuliaLang/julia/pull/31626#issuecomment-480319054) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/294aa1509dad8dfb64811f40f81ac383d72d27d8/f61bb2e_vs_3fbd7e5/report.md). cc @ararslan