@KristofferC 2020-02-02
I think this caused the doctests to start failing.